Remove API-specific Texture subclass headers and code duplication for loading & SetSize().
